requirements.txt
beautifulsoup4==4.14.2
boto3==1.40.72
botocore==1.40.72
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
jmespath==1.0.1
lxml==6.0.2
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
